test(ai): remove physical io assertions
